requirements.txt
Django==5.1.4
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.1
PyJWT==2.8.0
cryptography==41.0.5
Pillow==10.0.0
drf-yasg==1.21.7 